+++ src/frameworks/web/components/commons/Input.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import styled from 'styled-components'
 
-const SInput = styled.input`
+const S_Input = styled.input`
   display: inline-block;
   padding: 0 20px;
   height: 60px;
@@ -22,7 +22,7 @@ const Input: React.FC<IProps> = (props) => {
   const { type, name, placeholder, onChange, onKeyDown, value } = props
 
   return (
-    <SInput className={"input"} type={type} name={name} placeholder={placeholder} onChange={onChange} onKeyDown={onKeyDown} value={value} />
+    <S_Input className={"input"} type={type} name={name} placeholder={placeholder} onChange={onChange} onKeyDown={onKeyDown} value={value} />
   )
 }